- Continued SEEDS API Integration, can now pay with seeds and earn karma through the OASIS API! :)
- Improved ProviderManager.

- Improved OASISDNAManager, can now register all providers in AutoFailOver, AutoLoadBalance & AutoReplicate lists. Plus more...

- Added additional error handling.

- Improved AvatarManager & HolonManager.

- Fixed bugs in MongoDBOASIS Provider.

- Lots of improvements/refactorings.

- Fixed lots of bugs across the OASIS.

+++ NextGenSoftware.OASIS.API.ONODE.WebAPI/Models/SEEDS/SendInviteToJoinSeedsRequest/SendInviteToJoinSeedsRequestBase.cs
@@ -4,7 +4,7 @@ namespace NextGenSoftware.OASIS.API.ONODE.WebAPI.Models
 {
     public abstract class SendInviteToJoinSeedsRequestBase : SeedsKarmaBase
     {
-        public float TransferQuanitity { get; set; }
-        public float SowQuanitity { get; set; }
+        public int TransferQuanitity { get; set; }
+        public int SowQuanitity { get; set; }
     }
 }
